administration contract is being built

+++ contracts/administration/contract/src/db/domain_limit.rs
@@ -1,14 +1,12 @@
-use alloc::{ string::{ ToString, String }, format, fmt::format };
+use alloc::{ string::{ ToString, String }, format };
 use common_lib::{
 	constants::common_keys::AdministractionStoreKeys,
 	db::{ store::Store, traits::Storable },
 	enums::contracts_enum::ContractKind,
 };
 
-use crate::types::TResult;
-
 const DEFAULT_DOMAIN_CHARS_COUNT: u8 = 3u8;
-const DEFAULT_DOMAIN_LIST_LENGTH: u16 = 1000u16;
+const DEFAULT_DOMAIN_LIST_LENGTH: u32 = 1000;
 
 /**
  * 1. Domain chars min count
@@ -17,8 +15,8 @@ const DEFAULT_DOMAIN_LIST_LENGTH: u16 = 1000u16;
 pub(crate) trait DomainLimit {
 	fn get_chars_min_count(&self, extension: &str) -> u8;
 	fn set_chars_min_count(&self, extension: &str, count: u8) -> ();
-	fn get_listing_limit(&self, kind: ContractKind) -> u16;
-	fn set_listing_limit(&self, kind: ContractKind, limit: u16) -> ();
+	fn get_listing_limit(&self, kind: ContractKind) -> u32;
+	fn set_listing_limit(&self, kind: ContractKind, limit: u32) -> ();
 }
 
 fn to_domain_list_limit_key(kind: &ContractKind) -> String {
@@ -37,12 +35,12 @@ impl DomainLimit for Store {
 		self.set(&key, count);
 	}
 
-	fn get_listing_limit(&self, kind: ContractKind) -> u16 {
+	fn get_listing_limit(&self, kind: ContractKind) -> u32 {
 		let key = to_domain_list_limit_key(&kind);
-		self.get::<u16>(&key).unwrap_or(DEFAULT_DOMAIN_LIST_LENGTH)
+		self.get::<u32>(&key).unwrap_or(DEFAULT_DOMAIN_LIST_LENGTH)
 	}
 
-	fn set_listing_limit(&self, kind: ContractKind, limit: u16) -> () {
+	fn set_listing_limit(&self, kind: ContractKind, limit: u32) -> () {
 		let key = to_domain_list_limit_key(&kind);
 		self.set(&key, limit);
 	}
